refactor: split ai-assistant file
