+++ hlx_statics/blocks/ai-assistant/ai-assistant_chat-ui.js
@@ -1,0 +1,160 @@
+// @ts-check
+import { createTag } from "../../scripts/lib-adobeio";
+import { aiApiClient } from "./ai-assistant_api-client";
+import { handleUserQuery } from "./ai-assistant_chat-controller";
+
+import {
+  CHAT_BUTTON_LABEL_CLOSE,
+  CHAT_BUTTON_LABEL_MINIMIZE,
+  CHAT_BUTTON_LABEL_OPEN,
+  CHAT_WINDOW_ID,
+  CHAT_WINDOW_LABEL_ID,
+  ELEMENTS,
+  SEND_ICON_SRC,
+  STOP_ICON_SRC,
+} from "./ai-assistant_constants";
+
+/**
+ * Creates the AI avatar element
+ */
+export const createAiAvatar = () => {
+  return createTag("div", {
+    class: "chat-ai-avatar",
+    "aria-hidden": true,
+  });
+};
+/**
+ * Creates the chat window header
+ */
+export const createChatWindowHeader = () => {
+  const chatWindowHeader = createTag("header", { class: "chat-window-header" });
+  chatWindowHeader.appendChild(createAiAvatar());
+  const label = createTag("h2", {
+    class: "chat-window-label",
+    id: CHAT_WINDOW_LABEL_ID,
+  });
+  label.textContent = "Adobe Developer AI assistant";
+  const minimizeButton = createTag("button", {
+    class: "chat-window-minimize",
+    type: "button",
+    "aria-label": CHAT_BUTTON_LABEL_MINIMIZE,
+  });
+  const minimizeButtonIcon = createTag("img", {
+    src: "/hlx_statics/icons/chevron-down.svg",
+    alt: "",
+    "aria-hidden": true,
+  });
+  minimizeButton.appendChild(minimizeButtonIcon);
+
+  const closeButton = createTag("button", {
+    class: "chat-window-close",
+    type: "button",
+    "aria-label": CHAT_BUTTON_LABEL_CLOSE,
+  });
+  const closeButtonIcon = createTag("img", {
+    src: "/hlx_statics/icons/dismiss.svg",
+    alt: "",
+    "aria-hidden": true,
+  });
+  closeButton.appendChild(closeButtonIcon);
+
+  chatWindowHeader.appendChild(label);
+  chatWindowHeader.appendChild(minimizeButton);
+  chatWindowHeader.appendChild(closeButton);
+  ELEMENTS.CHAT_WINDOW_MINIMIZE_BUTTON = minimizeButton;
+  ELEMENTS.CHAT_WINDOW_CLOSE_BUTTON = closeButton;
+  return chatWindowHeader;
+};
+
+export const showStopButton = () => {
+  const btn = ELEMENTS.CHAT_SEND_BUTTON;
+  btn.querySelector("img").src = STOP_ICON_SRC;
+  btn.querySelector("span").textContent = "Stop response";
+  btn.classList.add("stop-mode");
+  btn.setAttribute("aria-label", "Stop response");
+  btn.disabled = false;
+};
+
+export const hideStopButton = () => {
+  const btn = ELEMENTS.CHAT_SEND_BUTTON;
+  btn.querySelector("img").src = SEND_ICON_SRC;
+  btn.querySelector("span").textContent = "";
+  btn.classList.remove("stop-mode");
+  btn.setAttribute("aria-label", "Send message");
+  btn.disabled = ELEMENTS.CHAT_TEXTAREA.value.trim() === "";
+};
+
+/**
+ * Creates the input section
+ */
+export const createInputSection = () => {
+  const inputSection = createTag("div", { class: "chat-window-input-section" });
+  const textarea = createTag("textarea", {
+    placeholder: "Type your message...",
+    rows: "4",
+    "aria-label": "Type your message",
+  });
+  const disclaimerText = createTag("div", { class: "chat-disclaimer-text" });
+  disclaimerText.innerHTML = `By using AI Assistant, you agree to the <a href="https://www.adobe.com/legal/licenses-terms/adobe-gen-ai-user-guidelines.html" target="_blank" rel="noopener noreferrer">Generative AI User Guidelines</a>.`;
+  const sendButton = createTag("button", {
+    class: "chat-send-button",
+    type: "button",
+    "aria-label": "Send message",
+  });
+  const sendButtonIcon = createTag("img", {
+    src: SEND_ICON_SRC,
+    alt: "",
+    "aria-hidden": true,
+    width: "20",
+    height: "20",
+  });
+  const sendButtonLabel = createTag("span", {});
+  sendButton.appendChild(sendButtonIcon);
+  sendButton.appendChild(sendButtonLabel);
+  sendButton.disabled = true;
+
+  const textareaWrapper = createTag("div", { class: "chat-textarea-wrapper" });
+  textareaWrapper.appendChild(textarea);
+  textareaWrapper.appendChild(sendButton);
+
+  inputSection.appendChild(textareaWrapper);
+  inputSection.appendChild(disclaimerText);
+  ELEMENTS.CHAT_SEND_BUTTON = sendButton;
+  ELEMENTS.CHAT_TEXTAREA = textarea;
+
+  sendButton.addEventListener("click", () => {
+    if (sendButton.classList.contains("stop-mode")) {
+      aiApiClient.abort();
+    } else {
+      handleUserQuery();
+    }
+  });
+  textarea.addEventListener("input", () => {
+    sendButton.disabled = textarea.value.trim() === "";
+  });
+  textarea.addEventListener("keydown", (e) => {
+    if (e.key === "Enter" && !e.shiftKey) {
+      e.preventDefault();
+      handleUserQuery();
+    }
+  });
+
+  return inputSection;
+};
+
+/**
+ * Creates the chat button
+ */
+export const createChatButton = () => {
+  const chatButton = createTag("button", {
+    class: "chat-button",
+    type: "button",
+    "aria-controls": CHAT_WINDOW_ID,
+    "aria-expanded": "false",
+    "aria-haspopup": "dialog",
+    "aria-label": CHAT_BUTTON_LABEL_OPEN,
+  });
+  chatButton.innerHTML = `<svg width="26" height="26" viewBox="0 0 26 26" focusable="false" aria-hidden="true" role="img" class="spectrum-Icon spectrum-Icon--sizeXL"><path d="M8.1248 24.6977C7.99531 24.6977 7.86582 24.6723 7.74267 24.6202C7.38339 24.4666 7.1498 24.1137 7.1498 23.7227V19.4977H6.1748C3.48721 19.4977 1.2998 17.3103 1.2998 14.6227V7.47266C1.2998 4.78506 3.48721 2.59766 6.1748 2.59766H11.3139C11.8521 2.59766 12.2889 3.03438 12.2889 3.57266C12.2889 4.11093 11.8521 4.54766 11.3139 4.54766H6.1748C4.56251 4.54766 3.2498 5.86036 3.2498 7.47266V14.6227C3.2498 16.235 4.56251 17.5477 6.1748 17.5477H8.1248C8.66308 17.5477 9.0998 17.9844 9.0998 18.5227V21.4299L12.8487 17.8206C13.0303 17.6454 13.2728 17.5477 13.5254 17.5477H19.8248C21.4371 17.5477 22.7498 16.235 22.7498 14.6227V12.9697C22.7498 12.4315 23.1865 11.9947 23.7248 11.9947C24.2631 11.9947 24.6998 12.4315 24.6998 12.9697V14.6227C24.6998 17.3103 22.5124 19.4977 19.8248 19.4977H13.9189L8.80147 24.4247C8.61611 24.6037 8.37236 24.6977 8.1248 24.6977Z" fill="#292929"/><path d="M17.2617 11.8078C17.0167 11.8078 16.7704 11.7443 16.5482 11.6161C16.0074 11.3051 15.7332 10.6868 15.8639 10.0774L16.4619 7.31493L14.5639 5.22147C14.145 4.75936 14.0726 4.08778 14.3837 3.54823C14.696 3.00868 15.3206 2.73446 15.9223 2.86395L18.6848 3.4619L20.7783 1.56395C21.2404 1.14627 21.9158 1.07517 22.4515 1.38368C22.9924 1.69472 23.2666 2.31297 23.1358 2.92234L22.5379 5.68484L24.4358 7.7783C24.8548 8.24041 24.9271 8.91199 24.6161 9.45154C24.3038 9.99237 23.6843 10.2691 23.0774 10.1358L20.3149 9.53788L18.2215 11.4358C17.9511 11.6808 17.6083 11.8078 17.2617 11.8078ZM17.0979 5.11355L18.0856 6.2028C18.3929 6.53796 18.5211 7.01022 18.4246 7.46218L18.1136 8.90184L19.2028 7.91414C19.5392 7.60691 20.0166 7.47996 20.4622 7.57517L21.9018 7.88621L20.9141 6.79696C20.6069 6.4618 20.4787 5.98954 20.5752 5.53758L20.8862 4.09793L19.797 5.08564C19.4618 5.39413 18.987 5.52362 18.5376 5.4246L17.0979 5.11355Z" fill="#292929"/><path d="M10.3125 14.9542C10.1449 14.9542 9.97734 14.911 9.82499 14.8234C9.45809 14.6114 9.2702 14.1874 9.35907 13.7735L9.59012 12.7071L8.8576 11.8997C8.57322 11.5861 8.5237 11.124 8.73572 10.7571C8.94774 10.3902 9.37429 10.2099 9.78563 10.2912L10.852 10.5222L11.6594 9.78972C11.9743 9.50535 12.4339 9.45583 12.802 9.66785C13.1689 9.87987 13.3568 10.3039 13.2679 10.7178L13.0369 11.7842L13.7694 12.5916C14.0538 12.9051 14.1033 13.3673 13.8913 13.7342C13.6793 14.1011 13.2515 14.2851 12.8414 14.2001L11.775 13.969L10.9676 14.7016C10.7835 14.8679 10.5486 14.9542 10.3125 14.9542Z" fill="#292929"/></svg>`;
+  ELEMENTS.CHAT_BUTTON = chatButton;
+  return chatButton;
+};
